requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.7
django-crispy-forms==2.1
django-etc==1.4.0
django-hitcount==1.3.5
Pillow==10.1.0
psycopg==3.1.13
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.8.0
